Add CONNECTIVITY_USE_RESTRICTED_NETWORKS permission to shell
The shell *already had* the privileges granted by this permission due to
specific code in netd, and it lost those privileges when it gained the
CHANGE_NETWORK_STATE permission. Explicitly add
CONNECTIVITY_USE_RESTRICTED_NETWORKS so that it can obtain sufficient
permission in netd no matter CHANGE_NETWORK_STATE is set or not.

Remove a duplicate row by the way.

Bug: 185071689
Test: atest NetdClientTest#protectFromVpnTcp6
Change-Id: I64bc321de2c83378ce7bc8d9eb3044ae7772faca

Former-commit-id: 81b55fbe8c5450fc37693ba609981d4bdeb5b671