Move bouncer to bottom of screen when in one handed mode.
If one handed bouncer is enabled, both in the SysUI config (i.e.
sw600dp-land) and in the device config, then use a different gravity for
the keyguard host view. This is used to move the bouncer to the bottom
of the screen when the bouncer is one handed.

Of course, password entry isn't one-handed, but we don't bother checking
which security mode is in use. In that mode, the password line is just
on top of the keyboard anyway, so even using gravity = BOTTOM, it
doesn't render differently.

Bug: 170858298
Test: atest
SystemUITests: com.android.keyguard.KeyguardHostViewControllerTest
Change-Id: Ia24d74ca1993d667c46b8ebb8bde6ea4888a0f97

Former-commit-id: 82f1652e350befb64cd3bcc6bdc6035b0da8c1ed