Use the existing wallpaper command API to send wake/sleep events to the wallpaper.
This is already used for similar events such as icon drops and user taps. Using it to indicate that the user has woken/slept the device (and passing along location information) makes sense!

This was previously rolled back since WallpaperManager can be null in Car environments. This CL now checks if WallpaperManager is null.

Bug: 169693662
Test: locally with modified wallpaper, it receives events
Change-Id: Ie9d21b4ff4a1edf871540481bc0c4216ed686b7f

Former-commit-id: d5da89479eafee5fb8e28da1b8f83eda32d44ddf