Dispatch pre-reboot verification tasks to the installer thread
Sometimes the system becomes busy and tasks dispatched to the background
thread are queued without running for a long time.

This causes test failures for the staged sessions didn't become ready
within the timeout. See b/185863896#comment2.

The issue is alleviated by using the installer thread instead of the
background thread. See comment#7 for the improvement.

Also, we have more consistency with this change by running all
installation tasks on the same thread for both staged and non-staged
sessions.

Bug: 185863896
Test: atest StagedRollbackTest MultiUserRollbackTest
Change-Id: Iec7e09defcfd1526bd84f71b97e0fa4ef734a37f

Former-commit-id: 80dff2f29abd35c0fed48812f1ad2c16e60738dd