Only allow the system to bind to the hotword detection service.
The hotword detection service is an isolated service which we want to
isolate from the rest of the system as much as possible. We don't want
to allow anybody but the system to bind it, not even its enclosing
application. For now special case the hotword detection service, until
we come up with a more generalized solution in T.

Bug: 181508704
Bug: 185746653
Test: local
Change-Id: Iea41383304e2632ff334a9719598db7510226c07

Former-commit-id: 7ced33fef8972c8d34358e3a08e5ecd667647f40