Clear mTranslucentActivityWaiting if the activity is going to relaunch
When client has request transition, the starting activity would convert
itself to be translucent, if there is a configuration change and
trigger the activity to relaunch, the activity will not become resumed
because the server thought its still translucent. So actually we do not
need to wait for translucent activity if it is going to relaunch.

Fixes: 184713003
Test: atest ActivityRecordTests
Change-Id: Ib66f89938672993610f59b6ad938300a309ccd87

Former-commit-id: 9e1c04ac30a2bffcb137b35880bb48bc9f71df51