Fix VIP conversations alerting incorrectly.
Transfer alerts (HUN/pulse) from the group to priority conversations which are isolated from that group.

Note: In cases where the alert is transferred from a group with 2+ children to the priority, there can be very small flickering because the group is inflated synchronously, so it starts displaying before the message is received to transfer the alert to the priority conversation, which then needs to be inflated asynchronously.

Fixes: 185680162
Test: Using WA, receive messages from 2+ people, where some but not all of those conversations have been set to Priority.  Test different numbers of people in each category (single vs. multile) and ensure that whenever an alert displays it shows the message that was just received.
Change-Id: I3c51cc20fe9ade4e24df2cc7d6784471be3ed509

Former-commit-id: 64c85ac6f5f20d031a7e192366b8335fab0253c2