Makes magnification switch UI never overlap navigationBar and statusBar
This change contrains the x,y postion of switch button to make
the button never overlap navigation bar window and status bar window.
When the orientation is changed or navigation mode is change, the
position of switch button should be ajusted by the draggable bounds
because the window insets are updated.

Bug: 185288922
Test: atest MagnificationModeSwitchTest; atest ModeSwitchesControllerTest;
Change-Id: Idc03275f39418c18a190cdff6ffb0e94280c6281

Former-commit-id: 732aed15d7312768378356e50b5fc80366b5ff85