Remap per app id for AR/location sources
Since the AR and location source concepts define what the
app can do, i.e. it can provide activity recognition or
location to other apps, we treat that app as the AR/location
source and allow op remapping for app users (as this app
runs for all users as a separate instance).

bug: 185504196

Test: manual

Change-Id: Iab15e2d4a2c3f6f15fe8a22c61ea7272460ea4d2

Former-commit-id: 4f0b921e9d6580001fb2f6d6542c285f51113a85