Synchronize PIP menu fade out with task dismiss
This CL addreses two issues:
1. PipMenuView.dismissPip was previously fading out the PIP
menu, waiting for that to complete, then notifying the task
organizer to remove and fade out the window, and then it would
fade the menu out again even though it was already hidden.
This resulted in a quick flash because the menu and the window
are not faded out in sync.
2. The menu fade out animation had a fixed duration of 150ms,
meaning that even if it starting fading out at the same time
as the PIP window, it would not finish at the same time because
the PIP exit anim duration is longer.

To fix both of them, the X button now hides the menu and starts
dismissing the task at the same time (once), and the dismiss
animation duration is equal to the task fade out animation
duration.

Previous: http://recall/-/e8RgaQ0dFlwFuh0c7GRpyy/g7XMbvBGMyAOKZMuFlYpW0
With CL: http://recall/-/e8RgaQ0dFlwFuh0c7GRpyy/g7XMbvBGMyAOKZMuFlYpW0

Bug: 182727923
Test: Tap X to dismiss PIP, no flashes seen
Change-Id: If9bdda8d0fd8b819c00ec4a592d1b2441d642d00

Former-commit-id: b2f5a2f98526977bc82fcc407bd06df20f002e5a