Add delay for mic/cam toggle dialog
This allows us to consolidate both mic and camera into a single dialog
when requested at approximately the same time. Also it allows (as a
workardound to b/181721787) for the dialog to be shown when an app
quickly opens another activity after opening mic/cam.

Also we special case telecom to show the microphone dialog when a phone
call is placed so that the user is able to easily take action on
unmuting the microphone.

Bug: 183985427
Fixes: 184163956
Test: atest CtsSensorPrivacyTestCases
Change-Id: I4cb897ee34d1ceeb4a2ad6104f812b15a2a4f357

Former-commit-id: 5c25e20c546693c4ac4b5fd0d1ea7d27c3183eff