Add leashes in launcher adapters in lieu of cross-process sync
Currently, SurfaceFlinger/BLAST doesn't provide any mechanism
for syncing across processes or any way to even know that a
particular transaction has been applied (exascerbated by the
fact that the transactions are queued up on the buffer queue).
This means there is always a race between when the last
launcher transaction is applied.

Until we get a solution, we add an extra leash so that launcher
doesn't actually modify the organized surface.

Bug: 186132704
Test: using winscope, observe that launcher animates leashes
      instead of the actual task surfaces
Change-Id: I53834fc434a33abd912ebd9edc06fc8c9bbf2e2f

Former-commit-id: e780af1e6f39e17bdf045ba15b91974729dd280c