Skip creating config context if it is the same as current
If the launching app is fullscreen on default display, the task
config should be the same as current global config. Then it is
unnecessary to create a context again.

Use diffPublicOnly because task config may have specified windowing
mode, which won't affect resources resolving.

It saves up to 4ms when adding splash screen starting window.
Also simply the initialization a bit.

Bug: 187104264
Test: OpenAppMicrobenchmark
Change-Id: I2bc7e8dfba3b1d8ad501f6935080a86f2c45f032

Former-commit-id: 0c54f2192f8f75745ecccd649e343a3c061c3d86