Allow network providers to set yield to bad wifi
While the "yield to bad wifi" policy is handled by CS for
backward compatibility, the network provider should in
fact be in control of it. There is no downside to allowing
providers to set it as no provider is using this, and it
will allow for migration in the future. The code is also
simpler, and becomes possible to test through CTS.

Test: NetworkScoreTest
Bug: 184037351
Change-Id: I3b2418f37bf1f0374d0a626fda7e2b6f3a3b41ba

Former-commit-id: c4660c98f6a63a73df1a79c34fbbf2266a6b381e