Fix sparkle alpha blending
We should not blend between the fill alpha and the sparkle alpha.
This would cause the sparkle pixel to have lower opacity than the fill,
when the fillAlpha > sparkleAlpha.

We do need to blend the fill and sparkle colors, but their alphas need
to be combined separately.

In the end we then premultiply the alpha into the color.

Test: manual
Bug: 186130682
Change-Id: I519f6c797839feda14e714c7375ee727aba2cf8b

Former-commit-id: a79111e05c0b28a48da88b0c30806ba0e84fa679