Log subject in ANR dump file for ANRs + watchdogs
This is needed to be able to have access to the Subject header in bug
reports and the ApplicationExitInfo API.

Adds the subject to the top of the ANR dump file. Where the dump
file is copied into the dropbox entry, the subject is no longer passed
in explicitly to the call to addErrorToDropBox to avoid including the
subject twice in the dropbox entry.

Verified e2e by triggering an ANR and confirming that the subject
appears at the top of the ANR file.

Test: atest FrameworksServicesTests
Bug: b/188122403
Change-Id: If2ee927a56e3f93c521307ca6355df7f560ea494

Former-commit-id: 8b8eb41e9b25197db9f2a69f0c19de7a0e62aa46