Lock screen a11y updates
When a11y features are enabled, inform the user
when the state of the lock screen changes.

Remove the window accessibility title on the udfps icon so that it won't
announce its window each time the udfps area receives focus. (ie:
previously was announcing "fingerprint sensor, SystemUI")

Add a label to the wallet affordance on the lock screen.

Test: manual
Fixes: 187139663
Fixes: 187190642
Fixes: 187082618
Fixes: 187080405
Change-Id: I4269b1943ec4d2e82eac9525947bb1903886f4f8

Former-commit-id: 1a21ef3fe1b80958430e0e2473142a48a6c5a7a9