Disables Smart AutoRotate in lock screen.
Currently camera based auto-rotate runs in the lock screen. It conflicts
with face auth and sometimes blocks face unlocking.

So this change will disable the smart auto-rotate whenever the key guard
is locked, and fall back to the recent screen orientation, or portrait
mode if there is no recent orientation.

Bug: 187429135
Test: atest WindowOrientationListenerTest
Test: manually tested the feature.

Change-Id: Idb38d60092de0fa7046e7b68baef9288805cb590

Former-commit-id: 6942cbb181fd7e26877a1a2584b7ec17ccd0a1db