Skip finishResize for non-exit animation if PIP is about to exit
This handles the case were exitPip is called while there is a
animateResize PIP ongoing, such the menu expand/unexpand animation.
When that was the case, the first finishResize was applied at the same
time as the exit WCT, resulting in bad frames at the end of the exit
animation.
This CL makes it so that the first finishResize is skipped if we
know exitPip is handling the exit and will eventually call its
own finishResize operation.

Bug: 185306679
Test: enter PIP with YT, tap on menu to show it and immediately
tap on the app launcher icon to expand PIP - verify the exit
transtion runs smoothly

Change-Id: Iac49f69cc698f1ac5d88fd408c204e2281b70ac7

Former-commit-id: 5e946e95aee09dd44c80f372a0277de3219e6525