Hook up nav-bar attach/detach logic to Shell Transitions
For the Recents gesture animation, the navbar gets attached to
the app while dragging. The actual navbar attach/detach logic
lives in Core, so this change basically hacks the logic into
the core-side of shell transitions. This is ugly and should
be a temporary solution until that behavior can move to Shell.

This also had to add an ugly API on IATM because the lifecycle
of the navbar reparent is different from the transition lifecycle
in this instance because of LiveTiles.

Bug: 188595497
Test: With shell transit enabled, open app, swipe up to recents.
      Observe navbar attached to app during drag and put back
      on screen afterwards. Also, no more crash.
Change-Id: I1aba53314329e3fd7678a7e8a5fc8ffa0294e395

Former-commit-id: 2996ef5f4f96d93e6397efce9551465dc9f10f68