[Injection] Stop using the component icon as a fallback option
Settings displays the component icon as a fallback option if the
metadata doesn't contain icon and icon_uri. Some injections that don't
need an icon have set a transpatrnt color to skip the fallback option.

However, the new Settings requires that an item without an icon should
align with the border. The transparent icon makes the item unaligned.

Therefore, we remove the hidden fallback option, and also filter out
the transparent color icon for backward compatibility.

Bug: 186801104
Test: robotest
Change-Id: I9221bba896e4785901ad617aa8ee58e6028ecc72

Former-commit-id: ce685fa47bd05998e3e9eba63d288dda7754fb0a