Update javadoc of TM#getDataNetworkType to align with implementation
Javadoc of TelephonyManager#getDataNetworkType says if the
TelephonyManager object is not created with createForSubscriptionId,
then applies to SubscriptionManager#getDefaultDataSubscriptionId().

But the actual implementation applies to
SubscriptionManager#getActiveDataSubscriptionId() instead. While they
are not always the same (e.g. when opptunistic network is active), the
implementation is more accurate than the javadoc.

Bug: 144803512
Test: make docs
Change-Id: I7db32a9a99dd5d7f7c32b6b33f3d5f7bd09a0e02

Former-commit-id: 0bacf4f46bc2ba465f81c352f394c12238e8ed14