Check for Ignore-AOSP-First commit tag
If the Ignore-AOSP-First tag is in a commit message, it means the
committer is intentionally bypassing the AOSP source of truth. Check
for this tag in the AOSP-first presubmit hook so that devs don't have
to turn off hooks to get past the AOSP-first policy in exceptional
cases.

Bug: 189139990
Change-Id: I18231b93d7f213c13ace0a0af24bb20a394cc50f
Merged-In: I18231b93d7f213c13ace0a0af24bb20a394cc50f
Test: Manual testing with various commit messages

Former-commit-id: f3d78421ba45ad5661f073795b3dc57411c2f09b