Provide a way to suppress page transitions.
Some pages are not suitable to apply the page transition, for example,
launching an acitivity that is depending on a value in the config file.

External apps that are not applying the same transition may be launched
in this case, applying the transition may lead to some unexpected
behaviour.

Hence, create a way to ignore the transition. Also refactor some logics.

Fixes: 189299479
Test: rebuild
Change-Id: Iebb47b9e0b0d9a3496f7e5f0c6cf1cd3d8d75031

Former-commit-id: abf096f21f7395202858c34eaa9f0f8195c1777d