sysui: show empty view in split mode
When showing the notification shade in split mode, the empty view was
not showing. This should fix it. If it is in split mode, it doesn't
matter if Quick Settings is expanded or not, it should always show the
empty view (depending on the other requisites).

Also changing the name of the test file and class so it is correctly
indexable in IntelliJ.

Bug: 188894961
Test: added test NotificationStackScrollLayoutControllerTest#testUpdateEmptyShadeView_splitShadeMode_alwaysShowEmptyView
Change-Id: I0d1228bff778cd9699afeda1431b69f77e3d3028

Former-commit-id: 4d49d726186315f4121edc8a8673b0f89b6031ce