Use sendBroadcastAsUser() instead of sendOrderedBroadcastAsUser()
Per offline discussion, we want to start the timer immediately after
sending the broadcast. We don't want to stall the install process for
whatever reason, no matter it is the delay from the broadcast or the
rollack manager.

This is also a mitigation of b/187384389 where sometimes
sendOrderedBroadcastAsUser() can be very slow.

Bug: 187383326
Test: atest StagedRollbackTest RollbackTest
Change-Id: I5314db8e0932088d39c3dffd1701ed1b87e446e0

Former-commit-id: d76dcf1a03db9d746088478894bbfd7393145594