Update wallet button size for large screens, reduce keyguard bottom text margin
The distortion of the wallet button occurs because keyguard
affordance width/height are overridden with different values
on larger screens (sw600dp) and width is greater than height.
Wallet button should have the same size on all screen sizes,
I created a separate pair of resources for this button.
Also updated keyguard bottom text margin, it should be lowered
so it is horizontally aligned with the wallet button and the
margin should be the same across all screen sizes.

Test: https://screenshot.googleplex.com/3BFdcFMWRLLv482.png (small screens)
Test: https://screenshot.googleplex.com/9d7oZrvAiwY7Vi2.png (large screens)
Bug: 189116030
Change-Id: Iac822e657e44d9664c616abe67096e48d594d7c0

Former-commit-id: df8edda1a2b262def4ce2b58c352255ce763b0b3