Add explicit versionCode to DexMetadataHelperTest APKs
These tests were verifying against an implicit versionCode 30 which was
added by the build system implicitly, using the old platform SDK. Since
one was not specified by the test apps.

When the SDK moved to 31, this changed what versionCode the APKs used,
and thus broke the tests. This uses 9001, which should be hard to
confuse with a near future platform SDK version.

Bug: 189990253

Test: atest DexMetadataHelperTest

Change-Id: Iaf39b1a7414e56e69a6eec57a339105e9e2ffbcc

Former-commit-id: 3c3532836bc1f8c06162b3af778493842cebda78