Allow app's targetSDK to overwrite SELINUX_LATEST_CHANGES SDK
The previous behavior capped the maximum target SDK level for apps to
the current platform target SDK in SELinux policy. This prevented
developing and testing newer SELinux policies in AOSP.

In addition to allowing higher targetSDKs specified by the app to be
considered in SELinux policies, this change adds a test to enforce that
behavior for future target SDK versions.

Bug: 190375530
Test: atest com.android.server.pm.SELinuxMMACTest
Test: atest CtsSelinuxTargetSdkCurrentTestCases
Test: atest CtsSelinuxTargetSdk29TestCases
Test: atest CtsSelinuxTargetSdk28TestCases
Test: atest CtsSelinuxTargetSdk27TestCases

Change-Id: Iee40e5ffd677038157ab7349a38eb34934458e25

Former-commit-id: bedbb8902f237dd54a928d1f53f30c5efea16bb6