Tuned the heads up animations more
The previous animation was using the linear interpolator
for the shape change, which is now properly using the
existing interpolator. Also the roundedness is now using
the interpolator to feel more connected to the rest of
the animation.
This also fixes some bugs where the heads up was animating
it's corner roundedness immediately, before it was
decided whether we will animating away, leading to some issues.

Lastly, this also adds a failsafe for the roundedness, which
could become really large if the roundness was updated without
an animation, while an animation was running already.

Fixes: 187175147
Test: add hun, observe nice animation
Change-Id: I4561eed97387e3465e7ea61066bffa11d8e6bda0

Former-commit-id: d2ad623032021187848a945c60bff8a830f5f092