Fix an issue where a custom permission is revoked
A normal permission declared by a system app could be revoked when
the updated system app is uninstalled. Assume that there are an apk
on /system and its updated apk on /data. And both declares the same
normal permission and it's granted to non-system apps. At this point,
If the updated system is uninstalled, That normal permission will be
revoked by uninstalling, but, it won't be granted to non-system apps
which had been had the permission when installing the apk on /system.
Because install permisisons were already fixed at install time.

This fix dosen't revoke normal permissions declared by the disabled
system app from apps while uninstalling the updated system app.

Bug: 182828078
Test: 1) Add a system app declares A, B permissions
      2) Update the system app declares A, B, C, D permissions
      3) Install a non-system app requesting A, B, C, D permissions
      4) Uninstall the updated system app
      5) Verify the non-system app has A, B permissions
Change-Id: I866934576dcc8103b123c10b0deb6144a85c67f5

Former-commit-id: c5d0555fedacbe1437c89c6648e465d6c2576733