Add low-power/thermal checks to highBrightnessMode
Adds an observer to HighBrightnessModeController to monitor the system's
thermal status and ensure HBM is off if the status is above the new
limit defined in the display-device-config file.

Also fixes a bug where AutomaticBrightnessController does not
automatically recalculate a value when HBM mode changes and the ambient
lux does not.

Bug: 179019497
Test: atest HighBrightnessModeControllerTest
Change-Id: I38a805671cee0f63e858b7ec0b6a45ca43df1737

Former-commit-id: 8f1a5c14ada36d551e18b950d6f3a41887d17bc7