Hook up the sysui backend for status bar content rects
StatusBarContentInsetsProvider provides the sysui backend for the
windowmanager api to communicate the location of the privacy indicators
to apps.

This CL hooks up that backend to the dot view controller and the
PhoneStatusBarView, which provides us with a stronger guarantee that
things are working and codifies the places where status bar content
layout can be changed. Also in this CL is a better dot positioning
algorithm that will ensure there is no flickering when the dot needs to
rotate.

Test: manual
Bug 187973222

Change-Id: I3dba7719ac4c4d178c9ec5265865961dbd0450f9
Merged-In: I3dba7719ac4c4d178c9ec5265865961dbd0450f9

Former-commit-id: 4a2c962e0da140a7d8dfd9b97c484923b99f0d7a