Fix positioning bug in user switcher
Introduces a new parameter to KeyguardVisibilityHelper to change whether
or not it will attempt to adjust the Y coordinates of its view.

Previously on the keyguard, if a PIN was set, the user switcher icon
would move upwards each time the phone was unlocked until it was
completely off the screen.

Test: With config_keyguardUserSwitcher=true, and
      config_keyguard_user_switch_opens_qs_details=false, set a screen
      PIN, repeatedly unlock and lock device, ensure that user switcher
      icon remains in the same place
Bug: 190351752
Change-Id: I4fec01dcc0ab39529ba2c7e74e4a5a13ac2d2cd8

Former-commit-id: 5b236efc8d95415437ace9443738972e69519afa