Fix not updating split bounds after devices unfolded
When the default display changed due to unfolded the device, the split
root tasks might be changed to invisible before reporting display info
changed to split component. This causing the split component didn't
update split layout due to the divider bar is not "initialized" after
released with stage roots changed to invisible.

This makes sure to update split layout while the split is activated no
matter the divider bar is initialized or not.

Fix: 189726605
Test: atest WMShellUnitTests
Test: check the split bounds updated properly after the device unfolded.
Change-Id: I722b08116a97b8482ea6e25b60ffe71005076cd7

Former-commit-id: 2af1cde04dde9313d5b39e858985fbb4d82937cb