Continue to dismiss keyguard when the remote animation is cancelled.
The keyguard going away process is consists of several steps between the
system and KeyguardService.

1. KeyguardService notifies the system that the Keyguard is going away.
2. The system prepares for the app transition, and notifies the
   KeyguardService that the app transition is ready.
3. The KeyguardService run a remote animation.
4. When the animation finishes, the KeyguardService notifies back to the
   system.

During the step 3, it's possible that the server cancels the animation
due to, for example, another app transition needs to be started. In this
case, we don't go back to the keyguard but finishes dismiss process.

Bug: 166736690
Bug: 188824028
Test: pass launcher3 MemoryTests, TaplTestsLauncher3 on test servers.
Change-Id: I3e45ca52b0006f89837365fb3f51a4372cc68060

Former-commit-id: b061c2bec95c6cb3968802307ff52ae0dfdea6b4