Bump x86 WebView zygote reservation size to 190MiB
The current 130MiB is not sufficient for the latest Chrome x86 dev build
of stripped native library. Official ARM library is around 59 MiB (on
disk). This flavor is slower to build and is less convenient to test
with.

Keep the 32bit arm reservation size unchanged, it is sufficient and
allows growth.

Bug: 184808875
Test: Manual: start a "release" build of Chrome/x86 (but not
      "official"), observe that with the change Chrome can
      mmap(MAP_FIXED) all the segments of the native library into the
      region successfully

Change-Id: I00507169eafbdcfc4d413bb9c76188e179ca0c0e

Former-commit-id: fac857dd313efe15098764b60ae6a5958be4304a