Modified the way the notifications clip to the boundary
Previously the notification shade didn't clip at all to
the qs - notification boundary, which could lead to
content peaking outside of the boundary during dismissal
or other situations. We now clip to the boundary
using rounded rect clipping. This also allows us to
remove various complex clipping logic that was updating
the clipping on the top for individual notification views.

This also fixes some bugs on the split shade where the
notifications were wrongly clipped previously.

This is change 2/2 to improve the notification dismissal.

Fixes: 181887261
Test: atest SystemUITests
Change-Id: I1ebe777c50b85a500ecce22823d547c6d9c3c6d5

Former-commit-id: 069b0c277a422d4ffa5f41d1150b7eaa4ee9c91c