Allow forcing status bar state changes and do so during a cancelled screen off.
During screen off, we show the AOD UI without fully switching to the KEYGUARD state. When screen off is cancelled, we ask all components to reset to the SHADE state, which should also reset the UI components we changed to show AOD. However, since the StatusBarState was already SHADE, this is ignored.

This adds a force flag, which we use when cancelling screen off to make sure that all UI components are reset to the SHADE state regardless.

Test: cancel screen off, pull down the shade
Fixes: 190986023
Change-Id: I79baeb71ac5d1ed45602ac55cdca996b3bed0ac3

Former-commit-id: b34e3b0cb7d0d5227e845a05fe58d3e286348a7a