Allow CacheOomRanker to rerank when there's <8 processes in the cache.
This allows us to increase the number of process we rerank without
losing the impact of the reranking.

This would also mean that the top apps in the cache get reranked more
often, so we also add a flag to prevent the top N processes from being
re-ranked.

Discussed with vvaldes@ offline: this is closer to the logic in the
simulations that decided the weights, so this logic change shouldn't
break any assumptions of the model.

Test: atest FrameworksMockingServicesTests:CacheOomRankerTest
Bug: 191357172
Change-Id: I0533da0a70cfe7ed1b48e4f98be24a46f17f7845

Former-commit-id: f913ae217a16b337a3125c2b49aad95a4946673e