Animate the Wallet activity launch on lockscreen (1/2)
This CL animates the launch of the Wallet activity from the wallet tile
or button on the lockscreen.

Note that doing so is still a bit flickery given that the wallpaper
disappears as soon as the animation starts. This will be fixed in a
follow-up CL.

See before/after videos in b/191438572#comment18.

Bug: 191438572
Test: Click Wallet tile or button on lockscreen (with card setup)
Change-Id: Ie2eaddf9e1ad7e620e1cf2260b8f25c6a8534109

Former-commit-id: 4179c7ce58fb4042445da5b420e7b64d4d4df278