Make screenshot animation display over cutout area
The screenshot preview is inset to avoid showing on top of the
cutout. However, the animation should display on top of the cutout,
since the screenshot is of the whole screen and includes the cutout.
This change shifts the start position of the animation by the
amount of the insets.

Bug: 191384262
Fix: 191384262
Test: manual
Change-Id: I2cfa0a7d9fec49b3c00f28f6127527b899cf131b

Former-commit-id: fcbf4c98003f7f49b83aea0772e81fa89ca5dd31