Capture a new scrim for scrolling screenshots
The capture process for scrolling screenshots is disorienting
(since the app itself is actually scrolled on the display), so we
add a scrim on top. Previously we just used the original bitmap,
but this looks odd if the user scrolled between the original
capture and tapping the scroll button.

This change captures a new screenshot when the scroll button is
tapped, and shades/dims it to indicate to the user that the
scrolling process is happening.

Bug: 183197533
Test: manual; visual verification
Change-Id: Ief2111ecc93f90638c6b2abb8a8120b83e805c07

Former-commit-id: a22342e8719ac5ca23221685adb62944f3489f1d