Add parsing class parcelling tests
Uses reflection to read all of the getters for each of the data classes
and finds their setters to input test data, serialize it, deserialize
it, and compare the before and after values.

Also records all of the methods available in the class, to ensure that
any new fields/methods are caught by the test.

Bug: 178852754

Test: atest com.android.server.pm.test.parsing.parcelling

Change-Id: I642fd43a3ff2e5675129f0f4ad3646c97e3f0e22

Former-commit-id: 593df2ce1ed128a61c6146bd3e98a049986d67d5