Changes default timeout for rotation memorization.
Rotation memorization is an optimization for smart-autorotate. It will
memorize last screen rotation before the screen turns off. And then use
that memorized rotation next time when the phone get unlocked.

The default timeout for the rotation memorization is 5 minutes. Based on
the droidfood experience, we decided to reduce it to 10 seconds.

Test: tested locally.
Bug: 192010410

Change-Id: Iad49eaa0399cfd4c02a014920b61a36aa171ce68

Former-commit-id: 51b7955cfaed63451be630a9a701d14136328be5