Properly reset FGS notification associations
When the service is brought down, the notification needs to be
properly freed from FGS-related policy.  When the same notification is
being shared by multiple live FGS instances, it needs to remain flagged
until the last of them exits that state.

Along the way, fix a latent bug that resulted in incorrect notification
behavior when an FGS's associated notification had id 0.

Bug: 191518901
Test: CtsAppTestCases:android.app.cts.ServiceTest
Change-Id: If8bae1c04d98cbab848823c86951544f792efba4

Former-commit-id: 0b3f85a600a41bf6920ce8ce7776d0fde2c13e29