Add onErrorCallback for TaskFragmentOrganizer
When app wants to create activity through sidecar (for example,
startActivityToSide), we need a callback in case it failed to finish the
request.

Allow the caller to set a token for the WCT, which will be passed back
if there is any error, so that the caller can identify the error
callback.

Bug: 190433129
Test: N/A
Change-Id: Icffb7255e75465b90ed822559296f147a7e3418a

Former-commit-id: ccff5a6a8b402e1069ac76c89b8467c46f098ae8