Clear invisible transition info
If 2 independent launch events happen in a short time, e.g. the
first activity is only added to the stack with initial states,
and only the latter is actually launched. Then the invisible
transition info (the activity occluded by the latter) should be
dropped. Otherwise when returning to it from back stack, the
transition time will still be counted from the time it added.

Bug: 192043869
Test: ActivityMetricsLaunchObserverTests#testOnReportFullyDrawn

Change-Id: I6c6ae59fce2f7e417c606489fb331d9ab6ad4f0a

Former-commit-id: 1314b9bd1a7925ea4c80e8889ca9318cd64dcb61