Fix side channel information disclosure
Application can query which packages are installed on the
device via the api PackageManager#setComponentEnabledSetting.
Two different exceptions are thrown from the api for the caller
to determine whether the application is installed.

Instead of throwing IllegalArgumentException at the beginning,
this cl throws SecurityException first if the caller tries to
modify others's state without the permission.

Bug: 184948501
Test: atest PackageManagerTest
Change-Id: Ie57a79c0c0fdc7bec7ae8aae5bbce48c3023878c

Former-commit-id: 82bebbe6ea32ed8ef66279477ff8d305fa9bd429