Allow apps to specify minimum chunk size.
JobScheduler wouldn't start a connectivity job if the app gave an estimated
download/upload size and JS calculated that the job wouldn't complete
successfully within the max execution time. The logic assumed that apps
don't support interruptible work and that downloads/uploads are
all-or-nothing. This negatively impacted apps that support
interruptible/resume downloads & uploads.

Adding an API to let an app indicate it supports resumable network
activity so that JS can be smarter about whether to start a job or not.

Bug: 188429037
Test: atest CtsJobSchedulerTestCases:JobInfoTest
Test: atest FrameworksMockingServicesTests:ConnectivityControllerTest
Change-Id: I76bfe8d1ad3ffedd7ce66e7e31098db089749c83

Former-commit-id: c14d4337d519e47121daf9ee4c20cece70061e8a