Don't fire pending remote input when is phone locked.
If an owner clicks "reply" on a locked device but doesn't
actually unlock their device, we shouldn't attempt to satisfy
the reply request. Check to ensure that the device is actually
unlocked before replying.

Also, if they click "emergency call", hide the shade before
showing the dialer, otherwise the shade can be left visible over
the top of the dialer.

Fixes: 191551623
Test: manual
Change-Id: Ie9e5127e19e222dcfa3c34b57aec1bc8074f5674

Former-commit-id: 9aae6eca53a6d00f22da761f41a84a838211043a