Update dynamic height adjusting
The current dynamic height adjusting workaround might cause that a
flicker problem happens in the collapsing toolbar and it might cause
that the collapsing toolbar has different background colors in some
setting pages. The new Material library has already supported the
dynamic height adjusting for multi-line titles, so we will use the new
mechanism instead.

Fix: 191038325
Fix: 189896657
Test: visual verified
1) Navigate to permission pages
2) Observe the title and check if it has different background color and
is overlapping with the back arrow

Change-Id: Ib928ab628cb7838460264f704bebf5c7a2aa9a47

Former-commit-id: 4bc5bdeb783eac126cb8db89c919821bf2c9cfa6