PropertyInvalidatedCache unit test
Bug: 192588767

This change corrects a latent defect related by the new bypass()
method in PropertyInvalidatedCache.  The defect is not current issue
because there are no clients using the bypass() method yet.

PropertyInvalidatedCache is modified to work without sytsem
properties, because test code generaly does not have permission to
read or write properties.  The modification has minimal impact when
there is not testing in progress.  New tests for basic cache
functionality are added.

Bypass statistics are included in 'dumpsys cacheinfo'.  We can
identify caches with a low hit ratio due to wildcard users, and
revisit the logic for those particular caches.

Lint warnings were corrected in the class javadoc comment and by
changing all occurrences of String.format() to
TextUtils.formatSimple().

Test: atest
 * FrameworksCoreTests:PropertyInvalidatedCacheTests

Change-Id: Ie3fbd2f4cb8a38b205a555af389814d7d64dd6bd

Former-commit-id: d28f1e4dece0848dc8bbcbb0f00972e5c16bbe05