Correct magnification switch button position after screen size is changed
This is for the foldable devices.
When the screen size is changed when the device is folded or  unfolded,
the magnification switch button position should be adjusted to keep the
X postion stick to the closest screen edge and Y position in the same
height ration.
https://buganizer.corp.google.com/issues/190365799#comment2

Bug: 190365799
Test: atest MagnificationModeSwitchTest
Mannual testing by using adb shell wm size 1080x2280 to change the screen size.

Change-Id: Ic73309c9d79fb8476b2de823f8d4d75355c4d944

Former-commit-id: 059dd66847d19d5b4347ae835fce070b309dda99