Enable matching of non-system app services
Change CurrentUserServiceSupplier to support attaching to services that
are not "system" services (i.e. provided by apps from the system image).
This is to support an upcoming test change in location time zone
detection where a fake provider supplied by a CTS app can be bound.

As part of this change constructors have been removed and replaced by
factory methods, which makes the choice to allow non-system services
more obvious.

Bug: 184947690
Test: build / treehugger
Change-Id: I629228c09ad1cf91c28c20d0b4b6227d1e4b1a00

Former-commit-id: 1a0b969d6a534627775f7f141ccf0a73b9b42aca