Remove hidden preconditions, binder usage in Vpn
The hidden APIs cause problems when Vpn is used in connectivity tests,
as their package (com.android.internal.util for Preconditions and
ThrowingRunnable used by lambdas) gets jarjared by Connectivity, but the
utils are not actually statically linked into the tests.

Remove usage of the hidden utilities, which will also be useful if Vpn
needs to build against SDK in the near future.

Bug: 187935317
Test: atest ConnectivityCoverageTests:ConnectivityServiceTest
Change-Id: I6152f07e7e88a88c202d78f03e370173fdd9a016

Former-commit-id: 053df0e78c71401bbe44eedc738569e7986e52c1