Withdraw suggestion when current user changes
This commit changes ControllerImpl so that it replaces previous
"certain" suggestions when the user changes. Before this commit, the
previous suggestion would be left until the providers made a new
suggestion. Generally, the ControllerImpl tries to avoid carrying state
between users so replacing the suggestion is a more consistent approach.

This commit also adds an assertion to an existing test. Before this
change, the test fails. It passes after.

It also adds a test for destroy(), which is another path where we expect
previous suggestions to be overridden, is potentially influenced by the
ControllerImpl change, and where test coverage was previously lacking.

Bug: 193410893
Test: atest com.android.server.timezonedetector.location
Change-Id: I9b1b61bdf0d7c269bd0e2252f8ba0663464eb9b8

Former-commit-id: 53196bbe6377b15de0a7fa65033044e228d09a71