Dovetail refresh for PowerMenu
Layout changes for each item in the PowerMenu including backgrounds,
padding and icons. This is to mimic Wear Material styling.

Also includes an update to the string for Emergency SOS action.

Later we might want to put some of these attributes as general themes to
that other alert dialogs in the framework can use the same styling

https://screenshot.googleplex.com/7xStDFauRa7Wcka

BUG: 173529191
Change-Id: I4e844186fc903f5366f491d0a1865a6a52ecd0fc

Former-commit-id: 35f887c1214a598903c14a1cfb6e787004ab0228