Rename RadioButtonPreference to SelectorWithWidgetPreference, and add a checkbox styling option.
This rename makes it slightly less confusing to have both radio and checkbox display options. Functionally, this preference was already equivalent to a checkbox (with checking/unchecking functionality managed by the caller); the reason to use this class for settings checkboxes rather than the base CheckBoxPreference is for the added widget functionality.

This change leaves RadioButtonPreference as-is to make it easier to bring google3 uses in without having to be in lockstep, but RadioButtonPreference will be deleted in a future CL.

Test: SelectorWithWidgetPreferenceTest, RunSettingsLibRoboTests
Bug: 190180868

Change-Id: I0811997b4625f1292b1842ea353a90221159bb50

Former-commit-id: 1516c28737ef99f526cfa846837305fbd0d1ea24